pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.1.1",
]
